Update chromeos ebuild to support running directly on target.
On the target there's no need to run qemu because we are already
on the right architecture.

BUG=chromium-os:15649
TEST=Verify chromeos package both emerges and gmerges fine

Reviewed-on: http://gerrit.chromium.org/gerrit/1374
Tested-by: David James <davidjames@chromium.org>

+++ chromeos-base/chromeos/chromeos-0.0.1.ebuild
@@ -239,9 +239,16 @@ qemu_run() {
 		*)
 			die "Unable to determine QEMU from ARCH."
 	esac
-	cp "/usr/bin/${qemu}" "${ROOT}/tmp" || die
-	chroot "${ROOT}" "/tmp/${qemu}" $* || die
-	rm "${ROOT}/tmp/${qemu}" || die
+
+	# If we're running directly on the target (e.g. gmerge), we don't need to
+	# chroot or use qemu.
+	if [ "${ROOT:-/}" == "/" ]; then
+		$* || die
+	else
+		cp "/usr/bin/${qemu}" "${ROOT}/tmp" || die
+		chroot "${ROOT}" "/tmp/${qemu}" $* || die
+		rm "${ROOT}/tmp/${qemu}" || die
+	fi
 }
 
 generate_font_cache() {
